internal/contour: move internal/listener into internal/contour
Updates #454

Move the contents of internal/listener nto internal/contour.
Deduplicate test and helpers.
