--- internal/listener/listener.go
@@ -1,168 +1,0 @@
-// Copyright © 2018 Heptio
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package listener
-
-import (
-	"sync"
-
-	"github.com/envoyproxy/go-control-plane/envoy/api/v2"
-	"github.com/gogo/protobuf/proto"
-)
-
-// ListenerCache manages the contents of the gRPC LDS cache.
-type ListenerCache struct {
-	// Envoy's HTTP (non TLS) listener address.
-	// If not set, defaults to DEFAULT_HTTP_LISTENER_ADDRESS.
-	HTTPAddress string
-
-	// Envoy's HTTP (non TLS) listener port.
-	// If not set, defaults to DEFAULT_HTTP_LISTENER_PORT.
-	HTTPPort int
-
-	// Envoy's HTTP (non TLS) access log path.
-	// If not set, defaults to DEFAULT_HTTP_ACCESS_LOG.
-	HTTPAccessLog string
-
-	// Envoy's HTTPS (TLS) listener address.
-	// If not set, defaults to DEFAULT_HTTPS_LISTENER_ADDRESS.
-	HTTPSAddress string
-
-	// Envoy's HTTPS (TLS) listener port.
-	// If not set, defaults to DEFAULT_HTTPS_LISTENER_PORT.
-	HTTPSPort int
-
-	// Envoy's HTTPS (TLS) access log path.
-	// If not set, defaults to DEFAULT_HTTPS_ACCESS_LOG.
-	HTTPSAccessLog string
-
-	// UseProxyProto configurs all listeners to expect a PROXY protocol
-	// V1 header on new connections.
-	// If not set, defaults to false.
-	UseProxyProto bool
-
-	cache
-}
-
-// httpAddress returns the port for the HTTP (non TLS)
-// listener or DEFAULT_HTTP_LISTENER_ADDRESS if not configured.
-func (lc *ListenerCache) httpAddress() string {
-	if lc.HTTPAddress != "" {
-		return lc.HTTPAddress
-	}
-	return DEFAULT_HTTP_LISTENER_ADDRESS
-}
-
-// httpPort returns the port for the HTTP (non TLS)
-// listener or DEFAULT_HTTP_LISTENER_PORT if not configured.
-func (lc *ListenerCache) httpPort() uint32 {
-	if lc.HTTPPort != 0 {
-		return uint32(lc.HTTPPort)
-	}
-	return DEFAULT_HTTP_LISTENER_PORT
-}
-
-// httpAccessLog returns the access log for the HTTP (non TLS)
-// listener or DEFAULT_HTTP_ACCESS_LOG if not configured.
-func (lc *ListenerCache) httpAccessLog() string {
-	if lc.HTTPAccessLog != "" {
-		return lc.HTTPAccessLog
-	}
-	return DEFAULT_HTTP_ACCESS_LOG
-}
-
-// httpsAddress returns the port for the HTTPS (TLS)
-// listener or DEFAULT_HTTPS_LISTENER_ADDRESS if not configured.
-func (lc *ListenerCache) httpsAddress() string {
-	if lc.HTTPSAddress != "" {
-		return lc.HTTPSAddress
-	}
-	return DEFAULT_HTTPS_LISTENER_ADDRESS
-}
-
-// httpsPort returns the port for the HTTPS (TLS) listener
-// or DEFAULT_HTTPS_LISTENER_PORT if not configured.
-func (lc *ListenerCache) httpsPort() uint32 {
-	if lc.HTTPSPort != 0 {
-		return uint32(lc.HTTPSPort)
-	}
-	return DEFAULT_HTTPS_LISTENER_PORT
-}
-
-// httpsAccessLog returns the access log for the HTTPS (TLS)
-// listener or DEFAULT_HTTPS_ACCESS_LOG if not configured.
-func (lc *ListenerCache) httpsAccessLog() string {
-	if lc.HTTPSAccessLog != "" {
-		return lc.HTTPSAccessLog
-	}
-	return DEFAULT_HTTPS_ACCESS_LOG
-}
-
-type cache struct {
-	mu      sync.Mutex
-	values  map[string]*v2.Listener
-	waiters []chan int
-	last    int
-}
-
-// Register registers ch to receive a value when Notify is called.
-// The value of last is the count of the times Notify has been called on this Cache.
-// It functions of a sequence counter, if the value of last supplied to Register
-// is less than the Cache's internal counter, then the caller has missed at least
-// one notification and will fire immediately.
-//
-// Sends by the broadcaster to ch must not block, therefor ch must have a capacity
-// of at least 1.
-func (c *cache) Register(ch chan int, last int) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	if last < c.last {
-		// notify this channel immediately
-		ch <- c.last
-		return
-	}
-	c.waiters = append(c.waiters, ch)
-}
-
-// Update replaces the contents of the cache with the supplied map.
-func (c *cache) Update(v map[string]*v2.Listener) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	c.values = v
-	c.notify()
-}
-
-// notify notifies all registered waiters that an event has occured.
-func (c *cache) notify() {
-	c.last++
-
-	for _, ch := range c.waiters {
-		ch <- c.last
-	}
-	c.waiters = c.waiters[:0]
-}
-
-// Values returns a slice of the value stored in the cache.
-func (c *cache) Values(filter func(string) bool) []proto.Message {
-	c.mu.Lock()
-	values := make([]proto.Message, 0, len(c.values))
-	for _, v := range c.values {
-		if filter(v.Name) {
-			values = append(values, v)
-		}
-	}
-	c.mu.Unlock()
-	return values
-}
